pyproject.toml
[project]
name = "latch-plots-faas"
version = "0.1.0"
description = "Latch Plots Runtime"
authors = [
    { name = "rteqs", email = "ryan@latch.bio" },
    { name = "maximsmol", email = "max@latch.bio" },
]
dependencies = [
    "latch-asgi>=1.0.1",
    "latch-o11y>=0.1.4",
    "uvloop>=0.19.0",
    "aiohttp>=3.9.3",
    "latch-postgres>=0.1.18",
    "numpy>=1.26.4",
    "plotly>=5.22.0",
    "psycopg<3.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "CC0-1.0" }

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = ["ruff>=0.5.5", "pandas>=2.2.2", "pyarrow>=17.0.0"]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
pydocstyle = { convention = "google" }
preview = true
extend-select = [
    "F",
    "E",
    "W",
    # "C90",
    "I",
    "N",
    "D",
    "UP",
    "YTT",
    "ANN",
    "ASYNC",
    "ASYNC1",
    "S",
    # "BLE", # `raise x from y` does not work
    "FBT",
    "B",
    "A",
    # "COM",
    # "CPY",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    # "EM",
    "EXE",
    "FA",
    "ISC",
    "ICN",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "ARG",
    "PTH",
    "TD",
    # "FIX",
    # "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "NPY",
    "AIR",
    "PERF",
    "FURB",
    "LOG",
    "RUF",
]
ignore = [
    "A002",
    "A003",

    "ANN101",
    "ANN102",

    "E402",
    "E501",
    "E741",

    "C408",
    "C901",

    "T201",

    "D415",

    "SIM108",

    "S101",
    "S603",

    "N807",

    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",

    "TRY003",
    "TRY300",
    "TRY301",

    "FLY002",

    "PLW2901",
    "PLW0603",

    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0914",
    "PLR0915",
    "PLR0916",
    "PLR0917",
    "PLR2004",

    "TD001",
    "TD003",
    "TD006",

    "TID252",

    "PD901",
]

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.pyright]
reportUnknownArgumentType = "none"
reportUnknownLambdaType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"

reportUnusedImport = "warning"
